using new train and test h5 db [seems good]

+++ cores/detector_wrapper.py
@@ -32,13 +32,13 @@ class DetectorWrapperV0(DetectorWrapperBase):
 
         # --- Configurable saving options ---
         self.svm_label_file = '/media/manu/ST8000DM004-2U91/tmp/afd'
-        self.save_as_svm = True
+        self.save_as_svm, self.save_as_h5 = False, True
         self.h5_path = '/media/manu/ST8000DM004-2U91/tmp/afd.h5'
 
         if self.save_as_svm and os.path.exists(self.svm_label_file):
             os.remove(self.svm_label_file)
 
-        if self.h5_path and os.path.exists(self.h5_path):
+        if self.save_as_h5 and self.h5_path and os.path.exists(self.h5_path):
             os.remove(self.h5_path)
 
     def _process_single(self, key, case_name, feat_sample=False):
@@ -77,7 +77,7 @@ class DetectorWrapperV0(DetectorWrapperBase):
         if self.save_as_svm:
             self.arc_detector.save_samples(path_save=self.svm_label_file)
 
-        if self.h5_path:
+        if self.save_as_h5:
             self.arc_detector.save_to_hdf5(path_save=self.h5_path)
 
         # self.arc_detector.save_seq()
